[Cinelog / Room] - migrate int ids to long - add tests to creators, and equals/hascode method to allow these tests to work - make creators return long for insertion methods => needed for crossref entities creation - add an id to Tmdb entity => used to avoid possible non unique primary key if serie and movie tmdb have a same id - add wishlist items migration

+++ app/src/main/java/com/ulicae/cinelog/data/dao/sqlite/DbReader.java
@@ -331,7 +331,7 @@ public class DbReader {
     private WishlistDataDto buildWishlistDto(Cursor cursor, WishlistItemType wishlistItemType, int biggestMovieReviewId) {
         return new WishlistDataDto(
                 cursor.getLong(0) + biggestMovieReviewId,
-                cursor.getInt(1),
+                cursor.getLong(1),
                 cursor.getString(2),
                 null,
                 null,
@@ -357,7 +357,7 @@ public class DbReader {
 
         return new WishlistDataDto(
                 cursor.getLong(0),
-                cursor.getInt(1),
+                cursor.getLong(1),
                 cursor.getString(2),
                 tmdbCursor.getString(1),
                 tmdbCursor.getString(2),
@@ -408,7 +408,7 @@ public class DbReader {
 
         return new WishlistDataDto(
                 cursor.getLong(0) + biggestMovieReviewId,
-                cursor.getInt(1),
+                cursor.getLong(1),
                 cursor.getString(2),
                 tmdbCursor.getString(1),
                 tmdbCursor.getString(2),
