+++ app/src/main/java/com/ulicae/cinelog/room/entities/WishlistTmdbCrossRef.java
@@ -5,10 +5,10 @@ import androidx.room.Entity;
 @Entity(primaryKeys = {"wishlistId", "tmdbId"})
 public class WishlistTmdbCrossRef {
 
-    public int wishlistId;
-    public int tmdbId;
+    public long wishlistId;
+    public long tmdbId;
 
-    public WishlistTmdbCrossRef(int wishlistId, int tmdbId) {
+    public WishlistTmdbCrossRef(long wishlistId, long tmdbId) {
         this.wishlistId = wishlistId;
         this.tmdbId = tmdbId;
     }
